- главный экран переделал согласно предоставляемым с сервера данным  - реализован вывод данных на экран  - сообщение о загрузке

+++ app/src/main/java/com/ponomarevss/myweatherapp/model/Wind.java
@@ -1,14 +1,14 @@
 package com.ponomarevss.myweatherapp.model;
 
 public class Wind {
-    private int speed;
+    private float speed;
     private int deg;
 
-    public int getSpeed() {
+    public float getSpeed() {
         return speed;
     }
 
-    public void setSpeed(int speed) {
+    public void setSpeed(float speed) {
         this.speed = speed;
     }
